pyproject.toml
[project]
name = "ta-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.27",
    "langchain-openai>=1.1.3",
    "langfuse>=3.10.6",
    "langgraph>=1.0.5",
    "mplfinance>=0.12.10b0",
    "openai>=2.11.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "loguru>=0.7.2",
    "langgraph-checkpoint-sqlite>=3.0.1",
]
